refactor: review 반영
- 컴포넌트 명 변경
- 변수명 변경
- 상수화
- 중복 코드 따로 빼내기
- json 파일 프로퍼티 수정
- 기타 코드 수정
