+++ frontend/src/components/result/ResultComponent.tsx
@@ -1,16 +1,9 @@
+import { css } from '@style/css';
+
 import { useCallback, useState } from 'react';
 
-import { css } from '../../../styled-system/css';
-import ResultDetailInfo from './ResultDetailInfo';
-import ResultUl from './ResultUl';
-
-const resultWrapperStyle = css({
-  padding: '24px',
-  minHeight: '40vh',
-  width: '80vw',
-  backgroundColor: 'darkgray',
-  margin: '0 auto',
-});
+import ResultList from './ResultList';
+import ResultDetailInfo from './ResultTotalInfo';
 
 export default function MswTestComponent() {
   const [isClicked, setIsClicked] = useState(false);
@@ -28,7 +21,7 @@ export default function MswTestComponent() {
         ) : (
           <>
             <p>채점 결과는 다음과 같습니다.</p>
-            <ResultUl testcaseNum={10} onSetIsAllTestDone={setIsAllTestDone} />
+            <ResultList testcaseNum={10} onSetIsAllTestDone={setIsAllTestDone} />
             <ResultDetailInfo isAllTestDone={isAllTestDone} />
           </>
         )}
@@ -37,3 +30,11 @@ export default function MswTestComponent() {
     </>
   );
 }
+
+const resultWrapperStyle = css({
+  padding: '24px',
+  minHeight: '40vh',
+  width: '80vw',
+  backgroundColor: 'darkgray',
+  margin: '0 auto',
+});
